Support aggressive caching with s3_include_abbrev_commit
CloudFront makes it difficult to update an existing URL with new
content.  In particular, they charge for invalidations, and often
require 15 minutes to process an invalidation request.

Similarly, aggressive HTTP caching makes it difficult to update the data
served by a URL.  We can address both of these issues by optionally
including a git commit hash in the asset path.

We choose to use the abbreviated commit (typically 7 characters) because
that is available on Heroku in the COMMIT_HASH environment variable.

Note that we support Cache-Control headers, but not Expires headers,
because the latter are not supported by the 's3' gem.

Doc for cache